+++ src/main/java/woop/logic/IllegalDescriptorException.java
@@ -1,4 +1,4 @@
-package WoopAI;
+package woop.logic;
 
 /**
  * Exception for when an input command does not fit the correct description needed to run.
