+++ src/main/java/woop/logic/UnknownCommandException.java
@@ -1,4 +1,4 @@
-package WoopAI;
+package woop.logic;
 
 /**
  * Exception for when a command is not recognised by the AI.
