+++ src/main/java/woop/logic/WoopLogic.java
@@ -1,6 +1,6 @@
-package WoopAI;
+package woop.logic;
 
-import ui.Ui;
+import woop.ui.Ui;
 
 /**
  * Handles main logic of Woop
