+++ src/main/java/woop/ui/DialogBox.java
@@ -1,4 +1,4 @@
-package ui;
+package woop.ui;
 
 import java.io.IOException;
 import java.util.Collections;
