+++ src/main/java/woop/ui/MainWindow.java
@@ -1,4 +1,4 @@
-package ui;
+package woop.ui;
 
 import javafx.application.Platform;
 import javafx.fxml.FXML;
@@ -8,7 +8,7 @@ import javafx.scene.control.TextField;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
 
-import WoopAI.WoopLogic;
+import woop.logic.WoopLogic;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.TimeUnit;
